Small refactorings to improve the structure
